# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EasyStack Backend API",
    "description": "A robust and scalable Node.js backend application with Express, TypeScript, and multi-database support (MongoDB and MySQL)",
    "version": "1.0.0",
    "contact": {
      "name": "Sanket More",
      "email": "contact@easystack.dev"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api-dev.easystack.dev",
      "description": "Development server"
    },
    {
      "url": "https://api-stage.easystack.dev",
      "description": "Staging server"
    },
    {
      "url": "https://api.easystack.dev",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "description": "Rotate refresh token and issue a new access token using refresh token cookie",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully (auth cookies updated)",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "1"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  }
                }
              },
              "verified": {
                "type": "boolean",
                "example": true
              },
              "message": {
                "type": "string",
                "example": "Email verified successfully"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Error message"
              },
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "statusCode": {
                "type": "integer",
                "example": 400
              }
            }
          }
        }
      },
...
    }
  }
}